+++ Data/PrimitiveArray/Index.hs
@@ -1,9 +1,10 @@
 
 module Data.PrimitiveArray.Index
   ( module Data.PrimitiveArray.Index.Class
-  , module Data.PrimitiveArray.Index.Complement
+--  , module Data.PrimitiveArray.Index.Complement
   , module Data.PrimitiveArray.Index.Int
-  , module Data.PrimitiveArray.Index.Outside
+  , module Data.PrimitiveArray.Index.IOC
+--  , module Data.PrimitiveArray.Index.Outside
   , module Data.PrimitiveArray.Index.PhantomInt
   , module Data.PrimitiveArray.Index.Point
   , module Data.PrimitiveArray.Index.Set
@@ -11,9 +12,10 @@ module Data.PrimitiveArray.Index
   ) where
 
 import Data.PrimitiveArray.Index.Class
-import Data.PrimitiveArray.Index.Complement
+--import Data.PrimitiveArray.Index.Complement
 import Data.PrimitiveArray.Index.Int
-import Data.PrimitiveArray.Index.Outside
+import Data.PrimitiveArray.Index.IOC
+--import Data.PrimitiveArray.Index.Outside
 import Data.PrimitiveArray.Index.PhantomInt
 import Data.PrimitiveArray.Index.Point
 import Data.PrimitiveArray.Index.Set