+++ Data/PrimitiveArray/Index/PhantomInt.hs
@@ -12,10 +12,15 @@ import Data.Ix(Ix)
 import Data.Serialize (Serialize)
 import Data.Typeable
 import Data.Vector.Fusion.Stream.Monadic (flatten,map,Step(..))
-import Data.Vector.Fusion.Stream.Size
 import Data.Vector.Unboxed.Deriving
 import GHC.Generics (Generic)
 import Prelude hiding (map)
+
+#if MIN_VERSION_vector(0,11,0)
+import Data.Vector.Fusion.Bundle.Size
+#else
+import Data.Vector.Fusion.Stream.Size
+#endif
 
 import Data.PrimitiveArray.Index.Class